+++ frontend/src/components/data-table/charts/__tests__/spec-snapshot.test.ts
@@ -1,10 +1,10 @@
 /* Copyright 2024 Marimo. All rights reserved. */
 
 import { describe, it, expect } from "vitest";
-import { createVegaSpec } from "../chart-spec";
-import { DEFAULT_BIN_VALUE, NONE_GROUP_BY } from "../chart-schemas";
+import { createVegaSpec } from "../spec";
+import { DEFAULT_BIN_VALUE, NONE_GROUP_BY } from "../schemas";
 import type { z } from "zod";
-import type { ChartSchema, ChartSchemaType } from "../chart-schemas";
+import type { ChartSchema, ChartSchemaType } from "../schemas";
 import { NONE_AGGREGATION, ChartType } from "../types";
 
 describe("createVegaSpec", () => {
