+++ frontend/src/components/data-table/charts/constants.ts
